vision_classification_project/api/requirements_api.txt
fastapi
uvicorn[standard]
python-multipart  # For handling file uploads
torch
torchvision
Pillow            # For image manipulation
numpy